requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.30.6
jinja2==3.1.6
python-multipart==0.0.22
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==46.0.6
paramiko==3.4.1
httpx==0.27.2
python-json-logger==2.0.7
apscheduler==3.10.4
netaddr==1.3.0
